+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.10.0] - 2021-03-01
+
+### Added
+
+-   data collection functionality
+-   doji corrections to prevent false flips
+
 ## [0.9.0] - 2021-03-01
 
 ### Added
@@ -196,6 +203,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 -   Working demo and API connections
 
+[0.10.0]: https://github.com/JesseChamberlain/binance-trader/compare/v0.9.0...v0.10.0
 [0.9.0]: https://github.com/JesseChamberlain/binance-trader/compare/v0.8.0...v0.9.0
 [0.8.0]: https://github.com/JesseChamberlain/binance-trader/compare/v0.7.0...v0.8.0
 [0.7.0]: https://github.com/JesseChamberlain/binance-trader/compare/v0.6.0...v0.7.0